requirements.txt
matrix-nio[e2e]==0.25.2
zeroc-ice==3.7.10.1
